chore(testdata): move test data handling to TestDataService OnModuleInit
