+++ backend/src/global/test-data/test-data.service.ts
@@ -1,11 +1,24 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, OnModuleInit } from '@nestjs/common';
+import { ActiveProfilesService } from '../active-profiles/active-profiles.service';
 import { UserTestDataService } from './user-test-data.service';
 
 @Injectable()
-export class TestDataService {
-  constructor(private readonly userTestDataService: UserTestDataService) {}
+export class TestDataService implements OnModuleInit {
+  constructor(
+    private readonly userTestDataService: UserTestDataService,
+    private readonly activeProfilesService: ActiveProfilesService
+  ) {}
 
-  public async resetTestData() {
-    await this.userTestDataService.onModuleInit();
+  public async onModuleInit() {
+    if (
+      this.activeProfilesService.isTestDataProfileActive() ||
+      this.activeProfilesService.isTestProfileActive()
+    ) {
+      await this.generateTestData();
+    }
+  }
+
+  public async generateTestData() {
+    await this.userTestDataService.generateTestData();
   }
 }
